+++ projects/ngx-highlightjs-demo/src/app/app.module.ts
@@ -6,8 +6,6 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { NgScrollbarModule } from 'ngx-scrollbar';
 import { MaterialModule } from './material/material.module';
 
-// import { HighlightPlusModule, GIST_OPTIONS } from '../../../ngx-highlightjs/plus/src/public_api';
-// import { HIGHLIGHT_OPTIONS } from '../../../ngx-highlightjs/src/public-api';
 import { HIGHLIGHT_OPTIONS } from 'ngx-highlightjs';
 import { HighlightPlusModule, GIST_OPTIONS } from 'ngx-highlightjs/plus';